refactor: move large inline test modules to sibling files
